% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_cmdv20.R
\name{calcpval}
\alias{calcpval}
\title{Computes the p-values for the canonical correlations}
\usage{
calcpval(OgX, OgY, X, Y, diag_W, finwgt, n_cc, rawcoef_var1, rawcoef_var2)
}
\arguments{
\item{OgX}{A matrix containing the first set of variables}

\item{OgY}{A matrix containing the second set of variables}

\item{X}{A matrix containing the standardized first set of variables}

\item{Y}{A matrix containing the standardized second set of variables}

\item{diag_W}{A matrix containing the diagonal of the survey weights}

\item{finwgt}{A vector containing the survey weights}

\item{n_cc}{Number of canonical correlations to be computed}

\item{rawcoef_var1}{A matrix containing the raw coefficients for the first set of variables}

\item{rawcoef_var2}{A matrix containing the raw coefficients for the second set of variables}
}
\value{
A list containing the p-values for the canonical correlations
}
\description{
Computes the p-values for the canonical correlations
}
