% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_cmdv20.R
\name{csdcanon}
\alias{csdcanon}
\title{Computes Canonical correlation p-values from regression coefficients}
\usage{
csdcanon(
  data,
  vars1,
  vars2,
  finwgt,
  n_cc,
  rawcoef_var1,
  rawcoef_var2,
  howmany = NA
)
}
\arguments{
\item{data}{A data frame containing the variables to be used in the canonical correlation analysis}

\item{vars1}{A vector of variable names for the first set of variables}

\item{vars2}{A vector of variable names for the second set of variables}

\item{finwgt}{Name of the column that contains the survey weights}

\item{n_cc}{Number of canonical correlations to be computed}

\item{rawcoef_var1}{A matrix containing the raw coefficients for the first set of variables}

\item{rawcoef_var2}{A matrix containing the raw coefficients for the second set of variables}

\item{howmany}{Number of canonical correlations to be computed}

\item{ccorr}{A matrix containing the canonical correlations}
}
\value{
A data frame with the p-values for the canonical correlations
}
\description{
Computes Canonical correlation p-values from regression coefficients
}
